refactor: 테스트 코드 함수명 한글로 재작성 수정
