+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,8 +1,6 @@
-package lotto;
+package lotto.domain;
 
 import java.util.stream.Stream;
-import lotto.domain.Lotto;
-import lotto.domain.Number;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
