+++ src/test/java/lotto/dto/LottoPrizesRecordTest.java
@@ -1,11 +1,12 @@
-package lotto.domain;
+package lotto.dto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.math.BigDecimal;
 import java.util.Map;
 import java.util.stream.Stream;
-import lotto.dto.LottoPrizesRecord;
+import lotto.domain.LottoPrice;
+import lotto.domain.LottoPrize;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
